add and remove top priority are plugged in
