+++ web/src/components/PriorityPicker/PriorityPickerUniversal/PriorityPickerUniversal.css
@@ -1,0 +1,6 @@
+.priority-picker-universal {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+}
