+++ web/src/components/PriorityPicker/PriorityPickerVote/PriorityPickerVote.css
@@ -1,0 +1,175 @@
+.priority-tabs {
+  /* display: flex;
+  flex-direction: column; */
+  margin-bottom: 10px;
+  /* justify-content: space-around; */
+  font-size: 14px;
+}
+
+.priority-tab {
+  padding: 10px;
+  font-family: 'gilroyextrabold', Helvetica, sans-serif;
+  cursor: pointer;
+  align-items: center;
+  color: #4d4d4d;
+}
+
+.my-vote-tab-wrapper {
+  padding: 0 24px;
+}
+
+.my-vote-tab-header {
+  height: 68px;
+}
+
+.priority-tab.active {
+  color: #5f65ff;
+  /* border-bottom: 3px solid #5f65ff; */
+}
+
+.priority-tab.active>.icon.very-small.enter-icon {
+  display: none;
+}
+
+.priority-tab-with-icon {
+  display: flex;
+}
+
+.icon.very-small.enter-icon {
+  margin-left: 6px;
+  display: block;
+}
+
+.icon.very-small.enter-icon .inner-icon {
+  background-color: #00a599;
+}
+
+.priority-tab.disabled {
+  opacity: 0.4;
+  cursor: default;
+}
+
+.aggregated-priority {
+  padding: 0 24px 22px 24px;
+}
+
+.my-vote, .aggregated-votes-wrapper {
+  -webkit-transition: transform 1s ease-in-out;
+  transition: transform 1s ease-in-out;
+  padding: 10px;
+}
+
+.my-vote {
+  padding: 10px;
+}
+
+.aggregated-votes-animation.closed {
+  transform: translateY(280px);
+}
+
+.aggregated-votes-animation.exit, .aggregated-votes-animation.enter-done {
+  transform: translateY(0);
+}
+
+.aggregated-votes-animation.exit-active {
+  transform: translateY(280px);
+  transition: transform 0.4s ease-out;
+}
+
+.aggregated-votes-animation.enter, .aggregated-votes-animation.exit-done {
+  transform: translateY(280px);
+}
+
+.aggregated-votes-animation.enter-active {
+  transform: translateY(0);
+  transition: transform 0.4s ease-in;
+}
+
+.my-vote-buttons {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 20px;
+}
+
+.remove-my-vote, .save-my-vote {
+  cursor: pointer;
+  font-family: 'gilroyextrabold', Helvetica, sans-serif;
+  color: #00a599;
+  transition: color 0.2s ease;
+  padding: 4px 0;
+}
+
+.remove-my-vote:hover, .save-my-vote:hover {
+  color: #6600ff;
+}
+
+.aggregated-priority-title {
+  font-family: 'hk-grotesk-medium', Helvetica, sans-serif;
+  margin-bottom: 10px;
+}
+
+.aggregated-priority-inputs {
+  color: #898989;
+  font-size: 12px;
+  padding: 0 10px 16px;
+}
+
+.priority-item {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+
+.priority-item-icon {
+  margin-right: 10px;
+}
+
+.priority-item-title {
+  width: 90px;
+  font-size: 14px;
+  font-family: 'hk-grotesk-medium', Helvetica, sans-serif;
+  color: #4d4d4d;
+}
+
+.priority-wrapper-button {
+  padding: 18px 40px;
+  text-align: center;
+  margin-bottom: 4px;
+}
+
+.priority-wrapper-footer .icon {
+  margin-right: 4px;
+}
+
+.my-vote-info {
+  padding: 0 10px 16px;
+  color: #898989;
+  font-size: 12px;
+}
+
+.aggregated-priority-wrapper {
+  z-index: 100;
+  position: relative;
+  background: #ffffff;
+  position: absolute;
+  top: 134px;
+  width: 100%;
+  padding: 6px 24px;
+  box-sizing: border-box;
+  border-top: 1px solid #e8e8e8;
+}
+
+.MuiSlider-thumb:hover {
+  box-shadow: 0px 0px 0px 8px rgba(255, 213, 203, 0.5) !important;
+}
+
+.MuiSlider-root.Mui-disabled {
+  color: #6600ff !important;
+}
+
+.MuiSlider-thumb.Mui-disabled {
+  display: none !important;
+  margin-top: -6px !important;
+  margin-left: -6px !important;
+  box-shadow: none !important;
+}
